requirements.txt
flask
flask-mysqldb
flask-cors
werkzeug
python-dotenv
google-generativeai
